get all users and user by id code refactoring
